:recycle: Move remote-check reconciliation into its own alpha module
The read-only remote/index reconciliation lived in the main repository module and
advertised a large stable-looking surface for what is still an alpha feature. Move
it into r3/remote_check.py as module-level functions taking the index and remotes
directly, and rename its result types with a leading underscore to signal they are
internal and provisional. Repository.remote_check stays as the thin public entry
point that delegates to r3.remote_check.check.

Also mark the whole remote backend module as alpha/provisional, and sharpen the
r3 remote check help and output so it is clear it reconciles the index against the
configured remote bucket(s) and does not verify local job integrity.
